+++ components/sidebar/SidebarLink.tsx
@@ -17,7 +17,8 @@ const SidebarLink = ({ link, Icon, text }: SidebarLinkProps): JSX.Element => {
     <Link
       href={link}
       className={`flex items-center space-x-3 cursor-pointer p-2 px-3 bg-transparent text-white hover:bg-[#1f1f1f] transition-all duration-100 ease-in-out ${
-        link.startsWith(pathname) && "border-t-2 lg:border-l-2 border-white !bg-[#393939]"
+        link.startsWith(pathname) &&
+        "border-t-2 lg:border-t-0 lg:border-l-2 border-white !bg-[#393939]"
       }`}
     >
       <Icon size={24} />