+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 07-08-2024
+
+Some improvements and changes.
+
+### Added
+
+- You can now join a project by using a join link now
+
+### Fixed
+
+- Sidebar looks normal now with only 1 border
+
+### Removed
+
+- Recordings page which never worked was removed
+
 ## [1.1.0] - 06-08-2024
 
 Finally, the full release! 🎉